+++ bench.py
@@ -125,6 +125,15 @@ def compress_multi_compress_to_buffer_list(chunks, opts, threads):
     zctx.multi_compress_to_buffer(chunks, threads=threads)
 
 
+@bench('discrete', 'stream_reader()')
+def compress_stream_reader(chunks, opts):
+    zctx = zstd.ZstdCompressor(**opts)
+    for chunk in chunks:
+        with zctx.stream_reader(chunk) as reader:
+            while reader.read(16384):
+                pass
+
+
 @bench('discrete', 'write_to()')
 def compress_write_to(chunks, opts):
     zctx = zstd.ZstdCompressor(**opts)